+++ src/components/del.vue
@@ -1,8 +1,8 @@
 <!--
  * @Author: 秦少卫
  * @Date: 2022-09-03 19:16:55
- * @LastEditors: June
- * @LastEditTime: 2023-05-19 22:08:43
+ * @LastEditors: 秦少卫
+ * @LastEditTime: 2023-06-20 13:26:20
  * @Description: 删除元素按钮
 -->
 
@@ -19,11 +19,12 @@ import { debounce } from 'lodash-es';
 const { canvas, mixinState } = useSelect();
 
 const del = debounce(function () {
-  const activeObject = canvas.c.getActiveObjects();
-  if (activeObject) {
-    activeObject.map((item) => canvas.c.remove(item));
-  }
-  canvas.c.requestRenderAll();
-  canvas.c.discardActiveObject();
+  canvas.editor.pluginEditor.del();
+  // const activeObject = canvas.c.getActiveObjects();
+  // if (activeObject) {
+  //   activeObject.map((item) => canvas.c.remove(item));
+  // }
+  // canvas.c.requestRenderAll();
+  // canvas.c.discardActiveObject();
 }, 300);
 </script>
